Reestrutura projeto conforme modelo simplificado
