+++ package.json
@@ -4,22 +4,11 @@
   "type": "module",
   "main": "main.js",
   "scripts": {
-    "build": "vite build",
-    "start": "cross-env NODE_ENV=production electron ."
+    "start": "electron ."
   },
   "dependencies": {
     "electron": "^37.3.0",
     "qrcode-terminal": "^0.12.0",
     "whatsapp-web.js": "^1.23.0"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-react": "^4.0.0",
-    "autoprefixer": "^10.4.16",
-    "cross-env": "^7.0.3",
-    "postcss": "^8.4.31",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "tailwindcss": "^3.4.1",
-    "vite": "^5.2.0"
   }
 }
