+++ renderer/style.css
@@ -1,0 +1,4 @@
+body {
+  font-family: Arial, sans-serif;
+  margin: 1rem;
+}
