--- src/App.jsx
@@ -1,8 +1,0 @@
-export default function App() {
-  return (
-    <div className="min-h-screen bg-gray-900 text-gray-100 p-4">
-      <h1 className="text-2xl mb-4">Extrator-I.A</h1>
-      <p>Aplicação Electron com React.</p>
-    </div>
-  );
-}
