--- tailwind.config.cjs
@@ -1,5 +1,0 @@
-module.exports = {
-  content: ['./index.html', './src/**/*.{js,jsx}'],
-  theme: { extend: {} },
-  plugins: []
-};
