--- vite.config.js
@@ -1,5 +1,0 @@
-import react from '@vitejs/plugin-react';
-
-export default {
-  plugins: [react()]
-};
