+++ whatsapp.js
@@ -14,14 +14,12 @@ if (!fs.existsSync(pastaDados)) fs.mkdirSync(pastaDados, { recursive: true });
 
 const sessoes = new Map();
 
-// Cria uma nova sessão e trata eventos básicos
 export function createSession(nome, janela) {
   const cliente = new Client({
     authStrategy: new LocalAuth({ clientId: nome })
   });
 
   cliente.on('qr', qr => {
-    // Exibe QR Code no console para autenticação
     qrcode.generate(qr, { small: true });
   });
 
@@ -33,7 +31,6 @@ export function createSession(nome, janela) {
     janela.webContents.send('session-status', { nome, status: 'offline' });
   });
 
-  // Salva mensagens recebidas em arquivos separados por número
   cliente.on('message', msg => {
     const numero = msg.from.replace('@c.us', '');
     const arquivo = path.join(pastaDados, `${numero}.json`);
@@ -51,17 +48,4 @@ export function createSession(nome, janela) {
 
   cliente.initialize();
   sessoes.set(nome, cliente);
-}
-
-// Exporta conversas em JSON ou TXT
-export function exportChat(numero, formato = 'json') {
-  const arquivoJson = path.join(pastaDados, `${numero}.json`);
-  if (!fs.existsSync(arquivoJson)) return;
-
-  if (formato === 'txt') {
-    const historico = JSON.parse(fs.readFileSync(arquivoJson));
-    const linhas = historico.map(m => `[${m.data}] ${m.de}: ${m.corpo}`).join('\n');
-    fs.writeFileSync(path.join(pastaDados, `${numero}.txt`), linhas);
-  }
-  // Se JSON, nada a fazer pois o arquivo já existe
 }
